+++ src/actions/initialize.ts
@@ -1,0 +1,4 @@
+export const ACTION_INITIALIZE = "@common/initialize" as const;
+export const initialize = () => ({
+  type: ACTION_INITIALIZE
+});