+++ src/index.tsx
@@ -5,8 +5,6 @@ import "@/style.css";
 
 import "@/debug";
 
-import "@/service-worker-init";
-
 import * as React from "react";
 import * as ReactDOM from "react-dom";