+++ src/services/offline-mode/reducer/index.ts
@@ -1,0 +1,9 @@
+import { reduceReducers } from "@/store/utils";
+
+import offlineProbeCompletedReducer from "./offline-probe-completed";
+import offlineSwitchCompletedReducer from "./offline-switch-completed";
+
+export default reduceReducers(
+  offlineProbeCompletedReducer,
+  offlineSwitchCompletedReducer
+);