+++ src/services/offline-mode/saga/index.ts
@@ -1,0 +1,9 @@
+import { fork } from "redux-saga/effects";
+
+import initializeSaga from "./initialize";
+import setOfflineEnabledSaga from "./set-offline-enabled";
+
+export default function* offlineModeSaga() {
+  yield fork(initializeSaga);
+  yield fork(setOfflineEnabledSaga);
+}