+++ src/state/index.ts
@@ -2,6 +2,10 @@ import { RouterState } from "connected-react-router";
 
 import { I18NState, defaultI18NState } from "@/services/i18n/state";
 import { OniSaveState, defaultOniSaveState } from "@/services/oni-save/state";
+import {
+  OfflineModeState,
+  defaultOfflineModeState
+} from "@/services/offline-mode/state";
 
 export * from "./utils";
 
@@ -10,6 +14,7 @@ export interface AppState {
   services: {
     i18n: I18NState;
     oniSave: OniSaveState;
+    offlineMode: OfflineModeState;
   };
 }
 
@@ -17,7 +22,8 @@ export const defaultAppState: Readonly<AppState> = {
   router: undefined as any,
   services: {
     i18n: defaultI18NState,
-    oniSave: defaultOniSaveState
+    oniSave: defaultOniSaveState,
+    offlineMode: defaultOfflineModeState
   }
 };
 Object.freeze(defaultAppState);