+++ src/store/reducer.ts
@@ -2,19 +2,20 @@ import { combineReducers, AnyAction } from "redux";
 
 import { connectRouter } from "connected-react-router";
 
-import i18nReducer from "@/services/i18n/reducer";
-import oniSaveReducer from "@/services/oni-save/reducer";
-import { OniSaveState, defaultOniSaveState } from "@/services/oni-save/state";
 import { AppState, defaultAppState } from "@/state";
 
+import i18nReducer from "@/services/i18n/reducer";
+import oniSaveReducer from "@/services/oni-save/reducer";
+import offlineModeReducer from "@/services/offline-mode/reducer";
+
 import history from "@/history";
+
 const routerReducer = connectRouter(history);
 
 const servicesReducer = combineReducers({
   i18n: i18nReducer,
-  // FIXME: Not sure why I need to do this, perhaps some reducer in the reduceReducered oniSaveReducer does not specify a default
-  oniSave: (state: OniSaveState = defaultOniSaveState, action: AnyAction) =>
-    oniSaveReducer(state, action)
+  oniSave: oniSaveReducer,
+  offlineMode: offlineModeReducer
 });
 
 export default function reducer(