+++ src/store/saga.ts
@@ -1,3 +1,9 @@
-import oniSaveSaga from "@/services/oni-save/saga";
+import { fork } from "redux-saga/effects";
 
-export default oniSaveSaga;
+import oniSaveSaga from "@/services/oni-save/saga";
+import offlineModeSaga from "@/services/offline-mode/saga";
+
+export default function* rootSaga() {
+  yield fork(oniSaveSaga);
+  yield fork(offlineModeSaga);
+}